+++ test/gst-msdk/encode/mpeg2.py
@@ -5,30 +5,11 @@
 ###
 
 from ....lib import *
-from ....lib.codecs import Codec
 from ....lib.gstreamer.msdk.util import *
-from ....lib.gstreamer.msdk.encoder import EncoderTest
+from ....lib.gstreamer.msdk.encoder import MPEG2EncoderTest
 
 spec      = load_test_spec("mpeg2", "encode")
 spec_r2r  = load_test_spec("mpeg2", "encode", "r2r")
-
-@slash.requires(*have_gst_element("msdkmpeg2enc"))
-@slash.requires(*have_gst_element("msdkmpeg2dec"))
-@slash.requires(*platform.have_caps("encode", "mpeg2"))
-class MPEG2EncoderTest(EncoderTest):
-  def before(self):
-    super().before()
-    vars(self).update(
-      caps          = platform.get_caps("encode", "mpeg2"),
-      codec         = Codec.MPEG2,
-      gstencoder    = "msdkmpeg2enc",
-      gstdecoder    = "msdkmpeg2dec",
-      gstmediatype  = "video/mpeg,mpegversion=2",
-      gstparser     = "mpegvideoparse",
-    )
-
-  def get_file_ext(self):
-    return "mpg"
 
 class cqp(MPEG2EncoderTest):
   def init(self, tspec, case, gop, bframes, qp, quality):
